Update citation information of preCICE
This removes the citing information for preCICE and instead links to citing information on the preCICE homepage. This avoids duplication of information. If the the citation information is updated on the homepage, we might forget to do update the README as well. Also restructure citation information to make it more verbose.